requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
numpy==1.26.2
Pillow>=10.2.0,<11.0.0
keyboard==0.13.5
pydirectinput